code refactoring

+++ benchmark_tests/benchmark_tests.h
@@ -10,7 +10,7 @@
 namespace Test::BenchMark {
 
     void perftDriver(int depth);
-    std::uint32_t perft(int depth, bool printInfo=true);
+    std::int64_t perft(int depth, bool printInfo=true);
     void standardPerft();
